+++ lib/sync/import.dart
@@ -4,11 +4,13 @@ import "dart:io";
 import "dart:typed_data";
 
 import "package:archive/archive_io.dart";
+import "package:csv/csv.dart";
 import "package:flow/entity/account.dart";
 import "package:flow/entity/category.dart";
 import "package:flow/entity/transaction.dart";
 import "package:flow/sync/exception.dart";
 import "package:flow/sync/import/base.dart";
+import "package:flow/sync/import/import_csv.dart";
 import "package:flow/sync/import/import_v1.dart";
 import "package:flow/sync/import/import_v2.dart";
 import "package:flow/sync/model/model_v1.dart";
@@ -41,6 +43,8 @@ Future<Importer> importBackup({File? backupFile}) async {
   final String ext = path.extension(file.path).toLowerCase();
 
   switch (ext.toLowerCase()) {
+    case ".csv":
+      return await _importCsv(file: file);
     case ".json":
       return await _importJson(file: file);
     case ".zip":
@@ -115,4 +119,15 @@ Future<Importer> _importJson({required File file}) async {
     2 => ImportV2(SyncModelV2.fromJson(parsed)),
     _ => throw UnimplementedError(),
   };
+}
+
+Future<Importer> _importCsv({required File file}) async {
+  final Stream<List<int>> readStream = file.openRead();
+  final List<List<dynamic>> rows =
+      await readStream
+          .transform(utf8.decoder)
+          .transform(CsvToListConverter())
+          .toList();
+
+  return ImportCSV(rows);
 }